feat: override default Int32 to Int64 scalar
This PR takes its roots from: https://github.com/dagger/graphql/commit/7ff025fae5140cd01097ab543efe0940ec10b0a9#diff-c97da2e4c49f2df9f25610f8474b7b2a69446d1256cd30bb5a114118b82d04e6.

We support JavaScript and its wacky numbers, seems like we'd need to limit ourselves to 9007199254740991: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/MAX_SAFE_INTEGER

The library was already taking into account the Int64 scalar, it was reverted to comply with `graphql-js`

+++ enum_type_test.go
@@ -14,13 +14,13 @@ var enumTypeTestColorType = graphql.NewEnum(graphql.EnumConfig{
 	Name: "Color",
 	Values: graphql.EnumValueConfigMap{
 		"RED": &graphql.EnumValueConfig{
-			Value: 0,
+			Value: int64(0),
 		},
 		"GREEN": &graphql.EnumValueConfig{
-			Value: 1,
+			Value: int64(1),
 		},
 		"BLUE": &graphql.EnumValueConfig{
-			Value: 2,
+			Value: int64(2),
 		},
 	},
 })
@@ -147,7 +147,7 @@ func TestTypeSystem_EnumValues_AcceptsEnumLiteralsAsInput(t *testing.T) {
 	query := "{ colorInt(fromEnum: GREEN) }"
 	expected := &graphql.Result{
 		Data: map[string]interface{}{
-			"colorInt": 1,
+			"colorInt": int64(1),
 		},
 	}
 	result := executeEnumTypeTest(t, query)
@@ -360,7 +360,7 @@ func TestTypeSystem_EnumValues_EnumValueMayHaveAnInternalValueOfZero(t *testing.
 	expected := &graphql.Result{
 		Data: map[string]interface{}{
 			"colorEnum": "RED",
-			"colorInt":  0,
+			"colorInt":  int64(0),
 		},
 	}
 	result := executeEnumTypeTest(t, query)
@@ -404,10 +404,10 @@ func TestTypeSystem_EnumValues_EnumValueMayBePointer(t *testing.T) {
 						},
 					}),
 					Resolve: func(_ graphql.ResolveParams) (interface{}, error) {
-						one := 1
+						one := int64(1)
 						return struct {
-							Color *int `graphql:"color"`
-							Foo   *int `graphql:"foo"`
+							Color *int64 `graphql:"color"`
+							Foo   *int64 `graphql:"foo"`
 						}{&one, &one}, nil
 					},
 				},
@@ -419,7 +419,7 @@ func TestTypeSystem_EnumValues_EnumValueMayBePointer(t *testing.T) {
 		Data: map[string]interface{}{
 			"query": map[string]interface{}{
 				"color": "GREEN",
-				"foo":   1}}}
+				"foo":   int64(1)}}}
 	result := g(t, graphql.Params{
 		Schema:        enumTypeTestSchema,
 		RequestString: query,
