+++ lists_test.go
@@ -65,7 +65,7 @@ func TestLists_ListOfNullableObjects_ContainsValues(t *testing.T) {
 		Data: map[string]interface{}{
 			"nest": map[string]interface{}{
 				"test": []interface{}{
-					1, 2,
+					int64(1), int64(2),
 				},
 			},
 		},
@@ -81,7 +81,7 @@ func TestLists_ListOfNullableObjects_ContainsNull(t *testing.T) {
 		Data: map[string]interface{}{
 			"nest": map[string]interface{}{
 				"test": []interface{}{
-					1, nil, 2,
+					int64(1), nil, int64(2),
 				},
 			},
 		},
@@ -115,7 +115,7 @@ func TestLists_ListOfNullableFunc_ContainsValues(t *testing.T) {
 		Data: map[string]interface{}{
 			"nest": map[string]interface{}{
 				"test": []interface{}{
-					1, 2,
+					int64(1), int64(2),
 				},
 			},
 		},
@@ -129,14 +129,14 @@ func TestLists_ListOfNullableFunc_ContainsNull(t *testing.T) {
 	// Note that its uses the expected signature `func() interface{} {...}`
 	data := func() interface{} {
 		return []interface{}{
-			1, nil, 2,
+			int64(1), nil, int64(2),
 		}
 	}
 	expected := &graphql.Result{
 		Data: map[string]interface{}{
 			"nest": map[string]interface{}{
 				"test": []interface{}{
-					1, nil, 2,
+					int64(1), nil, int64(2),
 				},
 			},
 		},
@@ -179,7 +179,7 @@ func TestLists_ListOfNullableArrayOfFuncContainsValues(t *testing.T) {
 		Data: map[string]interface{}{
 			"nest": map[string]interface{}{
 				"test": []interface{}{
-					1, 2,
+					int64(1), int64(2),
 				},
 			},
 		},
@@ -206,7 +206,7 @@ func TestLists_ListOfNullableArrayOfFuncContainsNulls(t *testing.T) {
 		Data: map[string]interface{}{
 			"nest": map[string]interface{}{
 				"test": []interface{}{
-					1, nil, 2,
+					int64(1), nil, int64(2),
 				},
 			},
 		},
@@ -224,7 +224,7 @@ func TestLists_NonNullListOfNullableObjectsContainsValues(t *testing.T) {
 		Data: map[string]interface{}{
 			"nest": map[string]interface{}{
 				"test": []interface{}{
-					1, 2,
+					int64(1), int64(2),
 				},
 			},
 		},
@@ -240,7 +240,7 @@ func TestLists_NonNullListOfNullableObjectsContainsNull(t *testing.T) {
 		Data: map[string]interface{}{
 			"nest": map[string]interface{}{
 				"test": []interface{}{
-					1, nil, 2,
+					int64(1), nil, int64(2),
 				},
 			},
 		},
@@ -287,7 +287,7 @@ func TestLists_NonNullListOfNullableFunc_ContainsValues(t *testing.T) {
 		Data: map[string]interface{}{
 			"nest": map[string]interface{}{
 				"test": []interface{}{
-					1, 2,
+					int64(1), int64(2),
 				},
 			},
 		},
@@ -308,7 +308,7 @@ func TestLists_NonNullListOfNullableFunc_ContainsNull(t *testing.T) {
 		Data: map[string]interface{}{
 			"nest": map[string]interface{}{
 				"test": []interface{}{
-					1, nil, 2,
+					int64(1), nil, int64(2),
 				},
 			},
 		},
@@ -364,7 +364,7 @@ func TestLists_NonNullListOfNullableArrayOfFunc_ContainsValues(t *testing.T) {
 		Data: map[string]interface{}{
 			"nest": map[string]interface{}{
 				"test": []interface{}{
-					1, 2,
+					int64(1), int64(2),
 				},
 			},
 		},
@@ -391,7 +391,7 @@ func TestLists_NonNullListOfNullableArrayOfFunc_ContainsNulls(t *testing.T) {
 		Data: map[string]interface{}{
 			"nest": map[string]interface{}{
 				"test": []interface{}{
-					1, nil, 2,
+					int64(1), nil, int64(2),
 				},
 			},
 		},
@@ -409,7 +409,7 @@ func TestLists_NullableListOfNonNullObjects_ContainsValues(t *testing.T) {
 		Data: map[string]interface{}{
 			"nest": map[string]interface{}{
 				"test": []interface{}{
-					1, 2,
+					int64(1), int64(2),
 				},
 			},
 		},
@@ -474,7 +474,7 @@ func TestLists_NullableListOfNonNullFunc_ContainsValues(t *testing.T) {
 		Data: map[string]interface{}{
 			"nest": map[string]interface{}{
 				"test": []interface{}{
-					1, 2,
+					int64(1), int64(2),
 				},
 			},
 		},
@@ -552,7 +552,7 @@ func TestLists_NullableListOfNonNullArrayOfFunc_ContainsValues(t *testing.T) {
 		Data: map[string]interface{}{
 			"nest": map[string]interface{}{
 				"test": []interface{}{
-					1, 2,
+					int64(1), int64(2),
 				},
 			},
 		},
@@ -618,7 +618,7 @@ func TestLists_NonNullListOfNonNullObjects_ContainsValues(t *testing.T) {
 		Data: map[string]interface{}{
 			"nest": map[string]interface{}{
 				"test": []interface{}{
-					1, 2,
+					int64(1), int64(2),
 				},
 			},
 		},
@@ -693,7 +693,7 @@ func TestLists_NonNullListOfNonNullFunc_ContainsValues(t *testing.T) {
 		Data: map[string]interface{}{
 			"nest": map[string]interface{}{
 				"test": []interface{}{
-					1, 2,
+					int64(1), int64(2),
 				},
 			},
 		},
@@ -782,7 +782,7 @@ func TestLists_NonNullListOfNonNullArrayOfFunc_ContainsValues(t *testing.T) {
 		Data: map[string]interface{}{
 			"nest": map[string]interface{}{
 				"test": []interface{}{
-					1, 2,
+					int64(1), int64(2),
 				},
 			},
 		},
@@ -873,7 +873,7 @@ func TestLists_ArrayOfNullableObjects_ContainsValues(t *testing.T) {
 		Data: map[string]interface{}{
 			"nest": map[string]interface{}{
 				"test": []interface{}{
-					1, 2,
+					int64(1), int64(2),
 				},
 			},
 		},
